+++ v1/quantum-card/src/main/java/com/ipeakoin/v1/quantum/card/dto/req/IncreaseBudgetBalanceReq.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
  */
 public class IncreaseBudgetBalanceReq {
     private String budgetId;
-    private BigDecimal cost;
+    private Integer cost;
     private String clientTransactionId;
 
     public String getBudgetId() {
@@ -21,11 +21,11 @@ public class IncreaseBudgetBalanceReq {
         this.budgetId = budgetId;
     }
 
-    public BigDecimal getCost() {
+    public Integer getCost() {
         return cost;
     }
 
-    public void setCost(BigDecimal cost) {
+    public void setCost(Integer cost) {
         this.cost = cost;
     }
